+++ examples/feature-sandbox/src/helloComponent.js
@@ -1,4 +1,4 @@
-import {Component, TemplateConfig, NgElement} from 'angular2/angular2.js';
+import {Component, Template, NgElement} from 'angular2/angular2.js';
 import ChildComponent from 'childComponent.js';
 import MyService from 'myService.js';
 import MyChildService from 'myChildService.js';
@@ -6,11 +6,11 @@ import MyChildService from 'myChildService.js';
 
 @Component({
     selector: '[hello-component]',
-    componentServices: [MyService, MyChildService, NgElement],
-    template: new TemplateConfig({
-        inline: '<h1>Hello {{ctrl.test}}</h1><child-component></child-component>',
-        directives: [ChildComponent]
-    })
+    componentServices: [MyService, MyChildService, NgElement]
+})
+@Template({
+    inline: '<h1>Hello {{ctrl.test}}</h1><child-component></child-component>',
+    directives: [ChildComponent]
 })
 class HelloComponent {
     constructor(test: MyService) {