+++ src/angular2/ng2/component.js
@@ -1,15 +1,10 @@
 import {Directive} from './directive.js';
-import {TemplateConfig} from './templateConfig.js';
 
 //Like a panel
 class Component extends Directive {
     constructor(options: ComponentOptions) {
         this.componentServices = options.componentServices;
-        this.template = options.template;
-
         delete options.componentServices;
-        delete options.template;
-
         super(options);
     }
 }
@@ -17,8 +12,7 @@ class Component extends Directive {
 var ComponentOptions = assert.define('ComponentOptions', function(options) {
     //Required fields
     assert(options).is(assert.structure({
-        selector: assert.string,
-        template: TemplateConfig
+        selector: assert.string
     }));
 
     //Optional fields