+++ src/angular2/ng2/template.js
@@ -1,9 +1,8 @@
-import {Directive} from './directive.js';
-
-//A transclude directive like ng-if or ng-repeat
-class Template extends Directive {
+class Template {
     constructor(options) {
-        super(options);
+        this.inline = options.inline;
+        this.url = options.url;
+        this.directives = options.directives;
     }
 }