+++ src/angular2/ng2/templateConfig.js
@@ -1,9 +1,0 @@
-class TemplateConfig {
-    constructor(options) {
-        this.inline = options.inline;
-        this.url = options.url;
-        this.directives = options.directives;
-    }
-}
-
-export {TemplateConfig};